+++ src/Command/Delete.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Phlib\Beanstalk\Command;
 
-use Phlib\Beanstalk\Connection\SocketInterface;
+use Phlib\Beanstalk\Connection\Socket;
 use Phlib\Beanstalk\Exception\CommandException;
 use Phlib\Beanstalk\Exception\NotFoundException;
 
@@ -25,7 +25,7 @@ class Delete implements CommandInterface
         return sprintf('delete %d', $this->id);
     }
 
-    public function process(SocketInterface $socket): self
+    public function process(Socket $socket): self
     {
         $socket->write($this->getCommand());