+++ src/Command/Kick.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Phlib\Beanstalk\Command;
 
-use Phlib\Beanstalk\Connection\SocketInterface;
+use Phlib\Beanstalk\Connection\Socket;
 use Phlib\Beanstalk\Exception\CommandException;
 
 /**
@@ -24,7 +24,7 @@ class Kick implements CommandInterface
         return sprintf('kick %d', $this->bound);
     }
 
-    public function process(SocketInterface $socket): int
+    public function process(Socket $socket): int
     {
         $socket->write($this->getCommand());
         $status = strtok($socket->read(), ' ');