+++ tests/Connection/SocketTest.php
@@ -6,34 +6,11 @@ namespace Phlib\Beanstalk\Connection;
 
 use Phlib\Beanstalk\Exception\SocketException;
 use phpmock\phpunit\PHPMock;
-use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
 class SocketTest extends TestCase
 {
     use PHPMock;
-
-    public function testImplementsInterface(): void
-    {
-        static::assertInstanceOf(SocketInterface::class, new Socket('localhost'));
-    }
-
-    public function testGetUniqueIdentifier(): void
-    {
-        $socket1 = new Socket('localhost', 11300);
-        $socket2 = new Socket('localhost', 11301);
-        static::assertNotSame($socket1->getUniqueIdentifier(), $socket2->getUniqueIdentifier());
-    }
-
-    public function testConnectOnSuccessReturnsSelf(): void
-    {
-        $fsockopen = $this->getFunctionMock(__NAMESPACE__, 'fsockopen');
-        $fsockopen->expects(static::any())->willReturn(true);
-        $stream_set_timeout = $this->getFunctionMock(__NAMESPACE__, 'stream_set_timeout');
-        $stream_set_timeout->expects(static::any())->willReturn(true);
-
-        static::assertInstanceOf(Socket::class, (new Socket('host'))->connect());
-    }
 
     public function testConnectOnFailureThrowsError(): void
     {
@@ -48,7 +25,7 @@ class SocketTest extends TestCase
         $stream_set_timeout = $this->getFunctionMock(__NAMESPACE__, 'stream_set_timeout');
         $stream_set_timeout->expects(static::any())->willReturn(true);
 
-        (new Socket('host'))->connect();
+        (new Socket('host'))->read();
     }
 
     public function testConnectsWithTheCorrectDetails(): void
@@ -70,9 +47,13 @@ class SocketTest extends TestCase
         $stream_set_timeout = $this->getFunctionMock(__NAMESPACE__, 'stream_set_timeout');
         $stream_set_timeout->expects(static::any())->willReturn(true);
 
+        // Add functional read after valid connection
+        $stream_get_line = $this->getFunctionMock(__NAMESPACE__, 'stream_get_line');
+        $stream_get_line->expects(static::any())->willReturn('');
+
         (new Socket($host, $port, [
             'timeout' => $timeout,
-        ]))->connect();
+        ]))->read();
     }
 
     public function testDisconnectWithValidConnection(): void
@@ -81,11 +62,17 @@ class SocketTest extends TestCase
         $fsockopen->expects(static::any())->willReturn(fopen('php://memory', 'r+'));
         $stream_set_timeout = $this->getFunctionMock(__NAMESPACE__, 'stream_set_timeout');
         $stream_set_timeout->expects(static::any())->willReturn(true);
-        $fsockopen = $this->getFunctionMock(__NAMESPACE__, 'fclose');
-        $fsockopen->expects(static::once())->willReturn(true); // <- test here
+
+        // Add functional read after valid connection
+        $stream_get_line = $this->getFunctionMock(__NAMESPACE__, 'stream_get_line');
+        $stream_get_line->expects(static::any())->willReturn('');
+
+        // Check the resource is closed
+        $fclose = $this->getFunctionMock(__NAMESPACE__, 'fclose');
+        $fclose->expects(static::once())->willReturn(true); // <- test here
 
         $socket = new Socket('host');
-        $socket->connect();
+        $socket->read();
         $socket->disconnect();
     }
 
@@ -95,11 +82,17 @@ class SocketTest extends TestCase
         $fsockopen->expects(static::any())->willReturn(true);
         $stream_set_timeout = $this->getFunctionMock(__NAMESPACE__, 'stream_set_timeout');
         $stream_set_timeout->expects(static::any())->willReturn(true);
-        $fsockopen = $this->getFunctionMock(__NAMESPACE__, 'fclose');
-        $fsockopen->expects(static::never()); // <- test here
+
+        // Add functional read to avoid error and allow testing disconnect()
+        $stream_get_line = $this->getFunctionMock(__NAMESPACE__, 'stream_get_line');
+        $stream_get_line->expects(static::any())->willReturn('');
+
+        // No attempt to close, as open didn't return a valid resource
+        $fclose = $this->getFunctionMock(__NAMESPACE__, 'fclose');
+        $fclose->expects(static::never());
 
         $socket = new Socket('host');
-        $socket->connect();
+        $socket->read();
         $socket->disconnect();
     }
 
@@ -117,7 +110,7 @@ class SocketTest extends TestCase
             )
             ->willReturn($dataLength);
 
-        $this->getMockSocket(['write'])
+        $this->getTestSocket()
             ->write($data);
     }
 
@@ -129,7 +122,7 @@ class SocketTest extends TestCase
         $fwrite->expects(static::any())
             ->willReturn(0);
 
-        $this->getMockSocket(['write'])
+        $this->getTestSocket()
             ->write('Some Data');
     }
 
@@ -138,7 +131,7 @@ class SocketTest extends TestCase
         $expectedData = 'Some Data';
         $stream_get_line = $this->getFunctionMock(__NAMESPACE__, 'stream_get_line');
         $stream_get_line->expects(static::any())->willReturn($expectedData);
-        static::assertSame($expectedData, $this->getMockSocket(['read'])->read());
+        static::assertSame($expectedData, $this->getTestSocket()->read());
     }
 
     public function testReadSuccessfullyWithLengthParam(): void
@@ -150,7 +143,7 @@ class SocketTest extends TestCase
         $fread = $this->getFunctionMock(__NAMESPACE__, 'fread');
         $fread->expects(static::any())->willReturn($expectedData);
 
-        static::assertSame($expectedData, $this->getMockSocket(['read'])->read(9));
+        static::assertSame($expectedData, $this->getTestSocket()->read(9));
     }
 
     public function testReadFailsWithBadData(): void
@@ -159,17 +152,31 @@ class SocketTest extends TestCase
 
         $stream_get_line = $this->getFunctionMock(__NAMESPACE__, 'stream_get_line');
         $stream_get_line->expects(static::any())->willReturn(false);
-        $this->getMockSocket(['read'])->read();
+        $this->getTestSocket()->read();
     }
 
-    /**
-     * @return Socket|MockObject
-     */
-    private function getMockSocket(array $originalMethods): Socket
+    public function testReadFailsWithBadDataWithLengthParam(): void
     {
-        $availableFns = ['connect', 'disconnect', 'read', 'write', 'getUniqueIdentifier'];
-        $mockedMethods = array_diff($availableFns, $originalMethods);
+        $this->expectException(SocketException::class);
 
-        return $this->createPartialMock(Socket::class, $mockedMethods);
+        $feof = $this->getFunctionMock(__NAMESPACE__, 'feof');
+        $feof->expects(static::any())->willReturn(false);
+        $fread = $this->getFunctionMock(__NAMESPACE__, 'fread');
+        $fread->expects(static::any())->willReturn(false);
+
+        $this->getTestSocket()->read(9);
+    }
+
+    private function getTestSocket(): Socket
+    {
+        $fsockopen = $this->getFunctionMock(__NAMESPACE__, 'fsockopen');
+        $fsockopen->expects(static::any())
+            ->willReturn(fopen('php://memory', 'r+'));
+
+        $stream_set_timeout = $this->getFunctionMock(__NAMESPACE__, 'stream_set_timeout');
+        $stream_set_timeout->expects(static::any())
+            ->willReturn(true);
+
+        return new Socket('host');
     }
 }