+++ tests/FactoryTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Phlib\Beanstalk;
 
-use Phlib\Beanstalk\Connection\ConnectionInterface;
 use Phlib\Beanstalk\Exception\InvalidArgumentException;
 use Phlib\Beanstalk\Pool\RandomStrategy;
 use Phlib\Beanstalk\Pool\RoundRobinStrategy;